py-openmeteo-api/requirements.txt
python-dotenv>=1.0
aio-pika>=6.8
httpx>=0.24